add main page content registration - 메인페이지 영화, tv card Javascript 를 통한 추가 - 목적에 맞는 js 파일명 및 함수명 변경
